requirements.txt
awsiotsdk==1.16.0
python-snap7
pymongo
configparser
bson